refactor(auth): restructure authentication feature into layered architecture (#10)
