+++ src/app/app.routes.ts
@@ -7,12 +7,12 @@ export const routes: Routes = [
   {
     path: 'register',
     loadComponent: () =>
-      import('./features/auth/register/register').then(m => m.RegisterComponent)
+      import('./features/auth/pages/register/register').then(m => m.RegisterComponent)
   },
   {
     path: 'login',
     loadComponent: () =>
-      import('./features/auth/login/login').then(m => m.LoginComponent)
+      import('./features/auth/pages/login/login').then(m => m.LoginComponent)
   },
   {
     path: '',
