+++ src/app/shared/layout/navigation/navigation.spec.ts
@@ -3,7 +3,7 @@ import { NavigationComponent } from './navigation';
 import { isSignal, signal } from '@angular/core';
 import { RouterTestingModule } from '@angular/router/testing';
 
-import { AuthService } from '../../../features/auth/services/auth-service/auth-service';
+import { AuthService } from '../../../features/auth/data-access/auth-service';
 
 class MockAuthService {
   isLogged = signal(false);
